refactor: step 관련 로직들을 모두 index.tsx로 옮김 그리고 내부 페이지 배열로 변경
