+++ src/pages/PostPages/WriteLayout/index.tsx
@@ -1,4 +1,3 @@
-import SafeArea from '@/common/components/SafeArea';
 import Header from './header';
 import Navigator from './navigator';
 import Step1 from '@/pages/PostPages/StepPages/step1';
@@ -12,31 +11,38 @@ import { useState } from 'react';
 
 import * as s from './style.css';
 
+const MAX_STEP = 6;
+
+const steps = [
+  <Step1 />,
+  <Step2 />,
+  <Step3 />,
+  <Step4 />,
+  <Step5 />,
+  <Step6 />,
+];
+
 const WriteLayout = () => {
+
   const [step, setStep] = useState(1);
 
-  const renderStep = () => {
-    switch (step) {
-      case 1:
-        return <Step1 />;
-      case 2:
-        return <Step2 />;
-      case 3:
-        return <Step3 />;
-      case 4:
-        return <Step4 />;
-      case 5:
-        return <Step5 />;
-      case 6:
-        return <Step6 />;
-    }
+  const isFirst = step === 1;
+  const isLast = step === MAX_STEP;
+
+  const goPrev = () => {
+    setStep((prev) => Math.max(1, prev - 1));  // 이전 상태 고려 (prev), 최소 1
+  }
+
+  const goNext = () => {
+    if (step < MAX_STEP) setStep(step + 1);
+    else console.log('작성 완료');
   };
 
   return (
     <div className={s.entireLayout}>
       <Header />
-      <div className={s.innerPage}>{renderStep()}</div>
-      <Navigator step={step} setStep={setStep} />
+      <div className={s.innerPage}>{steps[step - 1]}</div>
+      <Navigator goNext={goNext} goPrev={goPrev} isFirst={isFirst} isLast={isLast} />
     </div>
   );
 };
