+++ src/pages/PostPages/WriteLayout/navigator.tsx
@@ -1,20 +1,14 @@
 import Btn from '@/common/components/Button';
 import * as s from './style.css';
 
-const MAX_STEP = 6;
+interface NavigatorProps {
+  goNext: () => void;
+  goPrev: () => void;
+  isFirst: boolean;
+  isLast: boolean;
+}
 
-const Navigator = ({ step, setStep }: { step: number; setStep: (next: number) => void }) => {
-  const goPrev = () => {
-    if (step > 1) setStep(step - 1);
-  };
-
-  const goNext = () => {
-    if (step < MAX_STEP) setStep(step + 1);
-    else console.log('작성 완료');
-  };
-
-  const isFirst = step === 1;
-  const isLast = step === MAX_STEP;
+const Navigator = ({ goNext, goPrev, isFirst, isLast }: NavigatorProps) => {
 
   const label = isLast ? '완료' : '다음';
